v0.7.14.c.2 Add creation of a new order from the basket popup
When users clicks add to basket, the design is saved, assigned to their user and saved in the database as a new order (which then can be viewed in the basket)

+++ src/main/java/dev/hotdeals/treecreate/model/TreeOrder.java
@@ -10,7 +10,7 @@ public class TreeOrder
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "order_id")
-    private long orderId;
+    private int orderId;
 
     @Basic
     @Column(name = "amount")
@@ -28,12 +28,12 @@ public class TreeOrder
     {
     }
 
-    public long getOrderId()
+    public int getOrderId()
     {
         return orderId;
     }
 
-    public void setOrderId(long orderId)
+    public void setOrderId(int orderId)
     {
         this.orderId = orderId;
     }
